Simple DX11 sample, in VC10, VC11, Win32, x64.

+++ Platform/DirectX11/CreateEffectDX1x.cpp
@@ -634,8 +634,8 @@ HRESULT WINAPI D3DX11CreateEffectFromFileUtf8(std::string text_filename_utf8,D3D
 		hr=D3DX11CreateEffectFromMemory(binaryBlob->GetBufferPointer(),binaryBlob->GetBufferSize(),FXFlags,pDevice,ppEffect);
 		if(hr==S_OK)
 		{
-		fileLoader->Save(binaryBlob->GetBufferPointer(),binaryBlob->GetBufferSize(),binary_filename_utf8.c_str(),false);
-	}
+			fileLoader->Save(binaryBlob->GetBufferPointer(),(unsigned int)binaryBlob->GetBufferSize(),binary_filename_utf8.c_str(),false);
+		}
 	}
 	else
 	{
