+++ Platform/DirectX9/SimulSkyRenderer.cpp
@@ -67,7 +67,7 @@ SimulSkyRenderer::SimulSkyRenderer(simul::sky::SkyKeyframer *sk)
 void SimulSkyRenderer::SaveTextures(const char *base_filename)
 {
 	std::string filename_root(base_filename);
-	int pos=filename_root.find_last_of('.');
+	int pos=(int)filename_root.find_last_of('.');
 	std::string ext=filename_root.substr(pos,4);
 	filename_root=filename_root.substr(0,pos);
 	for(int i=0;i<3;i++)
